Fixing bug in rs_decode
The bug caused decoding errors for codes with symbol size 16 (and maybe
others). The bug was due to using the wrong data type for an
intermediate variable.

+++ src/reed_solomon.c
@@ -120,7 +120,7 @@ static inline void update_si(struct rs_code *rs, uint16_t *s, uint16_t data, int
 	if (s[i] == 0) {
 		s[i] = data;
 	} else {
-		uint16_t tmp = index_of[s[i]] + (fcr + i) * prim;
+		int tmp = index_of[s[i]] + (fcr + i) * prim;
 		s[i] = data ^ alpha_to[modnn(rs, tmp)];
 	}
 }
